Updates to compile with gcc8 on Ubuntu20 - add conditionals/entries in CMakeLists.txt for GFORTRAN, BLAS, LAPACK, QUADMATH _LIBRARIES and -fno-builtin for Fortran_FLAGS as math.h no longer defines powf_finite, exp2_finite and other symbols as of version 2.31 of the GNU C library. Cast malloc call to unsigned int in ./utils/mriio.cpp to prevent error.

+++ utils/mriio.cpp
@@ -431,7 +431,8 @@ int mriio_command_line(int argc, char *argv[])
   /* --- space for spaces and \0 --- */
   length += argc;
 
-  command_line = (char *)malloc(length);
+  /* command_line = (char *)malloc(length); */
+  command_line = (char *)malloc((unsigned int)length);
 
   c = command_line;
   for (i = 0; i < argc; i++) {
